migrate wakeup tests
