+++ clients/src/test/java/org/apache/kafka/clients/consumer/internals/AsyncKafkaConsumerTest.java
@@ -215,82 +215,6 @@ public class AsyncKafkaConsumerTest {
     }
 
     @Test
-    public void testWakeupBeforeCallingPoll() {
-        final String topicName = "foo";
-        final int partition = 3;
-        final TopicPartition tp = new TopicPartition(topicName, partition);
-        doReturn(Fetch.empty()).when(fetchCollector).collectFetch(any(FetchBuffer.class));
-        Map<TopicPartition, OffsetAndMetadata> offsets = mkMap(mkEntry(tp, new OffsetAndMetadata(1)));
-        doReturn(offsets).when(applicationEventHandler).addAndGet(any(FetchCommittedOffsetsApplicationEvent.class), any(Timer.class));
-        consumer.assign(singleton(tp));
-
-        consumer.wakeup();
-
-        assertThrows(WakeupException.class, () -> consumer.poll(Duration.ZERO));
-        assertDoesNotThrow(() -> consumer.poll(Duration.ZERO));
-    }
-
-    @Test
-    public void testWakeupAfterEmptyFetch() {
-        final String topicName = "foo";
-        final int partition = 3;
-        final TopicPartition tp = new TopicPartition(topicName, partition);
-        doAnswer(invocation -> {
-            consumer.wakeup();
-            return Fetch.empty();
-        }).when(fetchCollector).collectFetch(any(FetchBuffer.class));
-        Map<TopicPartition, OffsetAndMetadata> offsets = mkMap(mkEntry(tp, new OffsetAndMetadata(1)));
-        doReturn(offsets).when(applicationEventHandler).addAndGet(any(FetchCommittedOffsetsApplicationEvent.class), any(Timer.class));
-        consumer.assign(singleton(tp));
-
-        assertThrows(WakeupException.class, () -> consumer.poll(Duration.ofMinutes(1)));
-        assertDoesNotThrow(() -> consumer.poll(Duration.ZERO));
-    }
-
-    @Test
-    public void testWakeupAfterNonEmptyFetch() {
-        final String topicName = "foo";
-        final int partition = 3;
-        final TopicPartition tp = new TopicPartition(topicName, partition);
-        final List<ConsumerRecord<String, String>> records = asList(
-            new ConsumerRecord<>(topicName, partition, 2, "key1", "value1"),
-            new ConsumerRecord<>(topicName, partition, 3, "key2", "value2")
-        );
-        doAnswer(invocation -> {
-            consumer.wakeup();
-            return Fetch.forPartition(tp, records, true);
-        }).when(fetchCollector).collectFetch(Mockito.any(FetchBuffer.class));
-        Map<TopicPartition, OffsetAndMetadata> offsets = mkMap(mkEntry(tp, new OffsetAndMetadata(1)));
-        doReturn(offsets).when(applicationEventHandler).addAndGet(any(FetchCommittedOffsetsApplicationEvent.class), any(Timer.class));
-        consumer.assign(singleton(tp));
-
-        // since wakeup() is called when the non-empty fetch is returned the wakeup should be ignored
-        assertDoesNotThrow(() -> consumer.poll(Duration.ofMinutes(1)));
-        // the previously ignored wake-up should not be ignored in the next call
-        assertThrows(WakeupException.class, () -> consumer.poll(Duration.ZERO));
-    }
-
-    @Test
-    public void testClearWakeupTriggerAfterPoll() {
-        final String topicName = "foo";
-        final int partition = 3;
-        final TopicPartition tp = new TopicPartition(topicName, partition);
-        final List<ConsumerRecord<String, String>> records = asList(
-            new ConsumerRecord<>(topicName, partition, 2, "key1", "value1"),
-            new ConsumerRecord<>(topicName, partition, 3, "key2", "value2")
-        );
-        doReturn(Fetch.forPartition(tp, records, true))
-            .when(fetchCollector).collectFetch(any(FetchBuffer.class));
-        Map<TopicPartition, OffsetAndMetadata> offsets = mkMap(mkEntry(tp, new OffsetAndMetadata(1)));
-        doReturn(offsets).when(applicationEventHandler).addAndGet(any(FetchCommittedOffsetsApplicationEvent.class), any(Timer.class));
-        consumer.assign(singleton(tp));
-
-        consumer.poll(Duration.ZERO);
-
-        assertDoesNotThrow(() -> consumer.poll(Duration.ZERO));
-    }
-
-    @Test
     public void testEnsureCallbackExecutedByApplicationThread() {
         final String currentThread = Thread.currentThread().getName();
         ExecutorService backgroundExecutor = Executors.newSingleThreadExecutor();
